refactor(gateway): split provider stream state
