+++ exp/runtime/models/providers/streaming.py
@@ -9,10 +9,10 @@ from contextlib import suppress
 from dataclasses import dataclass
 from typing import cast
 
-from pydantic import JsonValue, ValidationError
+from pydantic import JsonValue
 
 from exp.common.core.artifacts import JsonObject
-from exp.common.models import ChatMaxTokensField, ToolCall
+from exp.common.models import ChatMaxTokensField
 from exp.runtime.gateway.contracts import (
     GatewayEvent,
     GatewayEventKind,
@@ -43,8 +43,11 @@ from exp.runtime.models.providers.streaming_events import (
     TERMINAL_EVENT_KINDS,
     OpenAIReasoningSummaryParser,
     ProviderOutputRetentionBudget,
+    ProviderToolAccumulator,
+    append_provider_tool_arguments,
     provider_refusal_failure,
-    retain_provider_entry,
+    require_provider_tool,
+    start_provider_tool_accumulator,
 )
 from exp.runtime.models.providers.streaming_requests import (
     anthropic_messages_stream_payload,
@@ -67,69 +70,6 @@ class _SseEvent:
 
     event: str | None
     data: str
-
-
-@dataclass
-class _ToolAccumulator:
-    """Provider-order state for one incrementally emitted function call."""
-
-    index: int
-    call_id: str
-    name: str
-    raw_arguments: str = ""
-    completed: bool = False
-
-    def complete(self) -> ToolCall:
-        """Parse the accumulated raw JSON once the provider closes the call.
-
-        Returns:
-            A complete tool call retaining provider-order argument text.
-
-        Raises:
-            ProviderResponseError: The raw arguments are not one JSON object.
-        """
-        try:
-            arguments = json.loads(self.raw_arguments)
-        except json.JSONDecodeError as exc:
-            raise ProviderResponseError("streamed tool arguments are not valid JSON") from exc
-        if not isinstance(arguments, dict):
-            raise ProviderResponseError("streamed tool arguments must decode to an object")
-        try:
-            return ToolCall(
-                call_id=self.call_id,
-                name=self.name,
-                arguments=arguments,
-                raw_arguments=self.raw_arguments,
-            )
-        except ValidationError as exc:
-            raise ProviderResponseError("streamed tool call is incomplete") from exc
-
-
-def _start_tool_accumulator(
-    tools: dict[int, _ToolAccumulator],
-    *,
-    index: int,
-    call_id: str,
-    name: str,
-    budget: ProviderOutputRetentionBudget,
-) -> _ToolAccumulator:
-    """Retain one provider tool identity under the shared output budget."""
-    budget.retain(call_id)
-    budget.retain(name)
-    tool = _ToolAccumulator(index=index, call_id=call_id, name=name)
-    retain_provider_entry(tools, index, tool)
-    return tool
-
-
-def _append_tool_arguments(
-    tool: _ToolAccumulator,
-    fragment: str,
-    *,
-    budget: ProviderOutputRetentionBudget,
-) -> None:
-    """Append one raw argument fragment only after reserving aggregate capacity."""
-    budget.retain(fragment)
-    tool.raw_arguments += fragment
 
 
 class _EventFactory:
@@ -570,7 +510,7 @@ async def _start_stream(
 async def _openai_responses_events(sse: _SseDecoder) -> AsyncIterator[GatewayEvent]:
     """Map native Responses lifecycle events to provider-neutral events."""
     factory = _EventFactory()
-    tools: dict[int, _ToolAccumulator] = {}
+    tools: dict[int, ProviderToolAccumulator] = {}
     retained_output = ProviderOutputRetentionBudget()
     reasoning_summaries = OpenAIReasoningSummaryParser(budget=retained_output)
     refusal_seen = False
@@ -604,7 +544,7 @@ async def _openai_responses_events(sse: _SseDecoder) -> AsyncIterator[GatewayEve
                     item.get("call_id") or item.get("id"), "OpenAI function call ID"
                 )
                 name = require_string(item.get("name"), "OpenAI function call name")
-                tool = _start_tool_accumulator(
+                tool = start_provider_tool_accumulator(
                     tools,
                     index=index,
                     call_id=call_id,
@@ -619,7 +559,7 @@ async def _openai_responses_events(sse: _SseDecoder) -> AsyncIterator[GatewayEve
                 )
                 initial = item.get("arguments")
                 if isinstance(initial, str) and initial:
-                    _append_tool_arguments(tool, initial, budget=retained_output)
+                    append_provider_tool_arguments(tool, initial, budget=retained_output)
                     yield factory.create(
                         GatewayEventKind.TOOL_ARGUMENTS_DELTA,
                         tool_call_index=index,
@@ -629,9 +569,9 @@ async def _openai_responses_events(sse: _SseDecoder) -> AsyncIterator[GatewayEve
                 raise ProviderResponseError("OpenAI stream emitted an unsupported output item")
         elif event_type == "response.function_call_arguments.delta":
             index = require_integer(payload.get("output_index"), "OpenAI output_index")
-            tool = _require_tool(tools, index)
+            tool = require_provider_tool(tools, index)
             delta = _optional_string(payload.get("delta"), "OpenAI argument delta")
-            _append_tool_arguments(tool, delta, budget=retained_output)
+            append_provider_tool_arguments(tool, delta, budget=retained_output)
             yield factory.create(
                 GatewayEventKind.TOOL_ARGUMENTS_DELTA,
                 tool_call_index=index,
@@ -655,7 +595,9 @@ async def _openai_responses_events(sse: _SseDecoder) -> AsyncIterator[GatewayEve
                             "OpenAI tool argument fragments changed at done"
                         )
                     if not tool.raw_arguments and final_arguments:
-                        _append_tool_arguments(tool, final_arguments, budget=retained_output)
+                        append_provider_tool_arguments(
+                            tool, final_arguments, budget=retained_output
+                        )
                         yield factory.create(
                             GatewayEventKind.TOOL_ARGUMENTS_DELTA,
                             tool_call_index=index,
@@ -723,7 +665,7 @@ async def _openai_responses_events(sse: _SseDecoder) -> AsyncIterator[GatewayEve
 async def _anthropic_messages_events(sse: _SseDecoder) -> AsyncIterator[GatewayEvent]:
     """Map native Anthropic Messages events to provider-neutral events."""
     factory = _EventFactory()
-    tools: dict[int, _ToolAccumulator] = {}
+    tools: dict[int, ProviderToolAccumulator] = {}
     retained_output = ProviderOutputRetentionBudget()
     input_tokens = 0
     output_tokens = 0
@@ -754,7 +696,7 @@ async def _anthropic_messages_events(sse: _SseDecoder) -> AsyncIterator[GatewayE
                 name = require_string(block.get("name"), "Anthropic tool name")
                 if index in tools:
                     raise ProviderResponseError("Anthropic stream repeated a tool-call start")
-                tool = _start_tool_accumulator(
+                tool = start_provider_tool_accumulator(
                     tools,
                     index=index,
                     call_id=call_id,
@@ -789,8 +731,8 @@ async def _anthropic_messages_events(sse: _SseDecoder) -> AsyncIterator[GatewayE
                     yield factory.create(GatewayEventKind.TEXT_DELTA, text_delta=text)
             elif delta_type == "input_json_delta":
                 fragment = _optional_string(delta.get("partial_json"), "Anthropic argument delta")
-                tool = _require_tool(tools, index)
-                _append_tool_arguments(tool, fragment, budget=retained_output)
+                tool = require_provider_tool(tools, index)
+                append_provider_tool_arguments(tool, fragment, budget=retained_output)
                 yield factory.create(
                     GatewayEventKind.TOOL_ARGUMENTS_DELTA,
                     tool_call_index=index,
@@ -867,7 +809,7 @@ async def _anthropic_messages_events(sse: _SseDecoder) -> AsyncIterator[GatewayE
 async def _openai_compatible_events(sse: _SseDecoder) -> AsyncIterator[GatewayEvent]:
     """Map generic Chat Completions chunks to provider-neutral events."""
     factory = _EventFactory()
-    tools: dict[int, _ToolAccumulator] = {}
+    tools: dict[int, ProviderToolAccumulator] = {}
     retained_output = ProviderOutputRetentionBudget()
     usage: GatewayUsage | None = None
     finish_reason: str | None = None
@@ -929,7 +871,7 @@ async def _openai_compatible_events(sse: _SseDecoder) -> AsyncIterator[GatewayEv
                 if tool is None:
                     call_id = require_string(item.get("id"), "OpenAI-compatible tool ID")
                     name = require_string(function.get("name"), "OpenAI-compatible tool name")
-                    tool = _start_tool_accumulator(
+                    tool = start_provider_tool_accumulator(
                         tools,
                         index=index,
                         call_id=call_id,
@@ -956,7 +898,7 @@ async def _openai_compatible_events(sse: _SseDecoder) -> AsyncIterator[GatewayEv
                 fragment = function.get("arguments")
                 if fragment is not None:
                     raw_fragment = _optional_string(fragment, "OpenAI-compatible argument delta")
-                    _append_tool_arguments(tool, raw_fragment, budget=retained_output)
+                    append_provider_tool_arguments(tool, raw_fragment, budget=retained_output)
                     yield factory.create(
                         GatewayEventKind.TOOL_ARGUMENTS_DELTA,
                         tool_call_index=index,
@@ -973,7 +915,7 @@ async def _openai_compatible_events(sse: _SseDecoder) -> AsyncIterator[GatewayEv
 
 async def _finish_open_tools(
     factory: _EventFactory,
-    tools: Mapping[int, _ToolAccumulator],
+    tools: Mapping[int, ProviderToolAccumulator],
 ) -> AsyncIterator[GatewayEvent]:
     """Complete every still-open tool call in provider index order."""
     for index in sorted(tools):
@@ -1015,17 +957,6 @@ def _optional_string(value: JsonValue | None, label: str) -> str:
     if not isinstance(value, str):
         raise ProviderResponseError(f"{label} must be text")
     return value
-
-
-def _require_tool(
-    tools: Mapping[int, _ToolAccumulator],
-    index: int,
-) -> _ToolAccumulator:
-    """Return one previously started provider tool call."""
-    try:
-        return tools[index]
-    except KeyError as exc:
-        raise ProviderResponseError("provider emitted arguments before a tool start") from exc
 
 
 async def _next_with_deadline[ValueT](
